Pull out gear into separate modules
