+++ common/script/content/gear/armor.js
