+++ common/script/content/gear/back.js
