+++ common/script/content/gear/gear-helper.js
@@ -1,0 +1,5 @@
+export function ownsItem (item) {
+  return (user) => {
+    return item && user.items.gear.owned[item];
+  };
+}
