+++ common/script/content/gear/head-accessory.js
