+++ common/script/content/gear/head.js
