+++ common/script/content/gear/sets/base.js
@@ -1,0 +1,80 @@
+import t from '../../translation';
+
+let armor = {
+  0: {
+    text: t('armorBase0Text'),
+    notes: t('armorBase0Notes'),
+    value: 0,
+  },
+};
+
+let back = {
+  0: {
+    text: t('backBase0Text'),
+    notes: t('backBase0Notes'),
+    value: 0,
+  },
+};
+
+let body = {
+  0: {
+    text: t('bodyBase0Text'),
+    notes: t('bodyBase0Notes'),
+    value: 0,
+  },
+};
+
+let eyewear = {
+  0: {
+    text: t('eyewearBase0Text'),
+    notes: t('eyewearBase0Notes'),
+    value: 0,
+    last: true,
+  },
+};
+
+let head = {
+  0: {
+    text: t('headBase0Text'),
+    notes: t('headBase0Notes'),
+    value: 0,
+  },
+};
+
+let headAccessory = {
+  0: {
+    text: t('headAccessoryBase0Text'),
+    notes: t('headAccessoryBase0Notes'),
+    value: 0,
+    last: true,
+  },
+};
+
+let shield = {
+  0: {
+    text: t('shieldBase0Text'),
+    notes: t('shieldBase0Notes'),
+    value: 0,
+  },
+};
+
+let weapon = {
+  0: {
+    text: t('weaponBase0Text'),
+    notes: t('weaponBase0Notes'),
+    value: 0,
+  },
+};
+
+let baseSet = {
+  armor,
+  back,
+  body,
+  eyewear,
+  head,
+  headAccessory,
+  shield,
+  weapon,
+};
+
+export default baseSet;
