+++ common/script/content/gear/sets/mystery.js
@@ -1,0 +1,340 @@
+import t from '../../translation';
+
+let armor = {
+  201402: {
+    text: t('armorMystery201402Text'),
+    notes: t('armorMystery201402Notes'),
+    mystery: '201402',
+    value: 0,
+  },
+  201403: {
+    text: t('armorMystery201403Text'),
+    notes: t('armorMystery201403Notes'),
+    mystery: '201403',
+    value: 0,
+  },
+  201405: {
+    text: t('armorMystery201405Text'),
+    notes: t('armorMystery201405Notes'),
+    mystery: '201405',
+    value: 0,
+  },
+  201406: {
+    text: t('armorMystery201406Text'),
+    notes: t('armorMystery201406Notes'),
+    mystery: '201406',
+    value: 0,
+  },
+  201407: {
+    text: t('armorMystery201407Text'),
+    notes: t('armorMystery201407Notes'),
+    mystery: '201407',
+    value: 0,
+  },
+  201408: {
+    text: t('armorMystery201408Text'),
+    notes: t('armorMystery201408Notes'),
+    mystery: '201408',
+    value: 0,
+  },
+  201409: {
+    text: t('armorMystery201409Text'),
+    notes: t('armorMystery201409Notes'),
+    mystery: '201409',
+    value: 0,
+  },
+  201410: {
+    text: t('armorMystery201410Text'),
+    notes: t('armorMystery201410Notes'),
+    mystery: '201410',
+    value: 0,
+  },
+  201412: {
+    text: t('armorMystery201412Text'),
+    notes: t('armorMystery201412Notes'),
+    mystery: '201412',
+    value: 0,
+  },
+  201501: {
+    text: t('armorMystery201501Text'),
+    notes: t('armorMystery201501Notes'),
+    mystery: '201501',
+    value: 0,
+  },
+  201503: {
+    text: t('armorMystery201503Text'),
+    notes: t('armorMystery201503Notes'),
+    mystery: '201503',
+    value: 0,
+  },
+  201504: {
+    text: t('armorMystery201504Text'),
+    notes: t('armorMystery201504Notes'),
+    mystery: '201504',
+    value: 0,
+  },
+  201506: {
+    text: t('armorMystery201506Text'),
+    notes: t('armorMystery201506Notes'),
+    mystery: '201506',
+    value: 0,
+  },
+  201508: {
+    text: t('armorMystery201508Text'),
+    notes: t('armorMystery201508Notes'),
+    mystery: '201508',
+    value: 0,
+  },
+  201509: {
+    text: t('armorMystery201509Text'),
+    notes: t('armorMystery201509Notes'),
+    mystery: '201509',
+    value: 0,
+  },
+  301404: {
+    text: t('armorMystery301404Text'),
+    notes: t('armorMystery301404Notes'),
+    mystery: '301404',
+    value: 0,
+  },
+};
+
+let back = {
+  201402: {
+    text: t('backMystery201402Text'),
+    notes: t('backMystery201402Notes'),
+    mystery: '201402',
+    value: 0,
+  },
+  201404: {
+    text: t('backMystery201404Text'),
+    notes: t('backMystery201404Notes'),
+    mystery: '201404',
+    value: 0,
+  },
+  201410: {
+    text: t('backMystery201410Text'),
+    notes: t('backMystery201410Notes'),
+    mystery: '201410',
+    value: 0,
+  },
+  201504: {
+    text: t('backMystery201504Text'),
+    notes: t('backMystery201504Notes'),
+    mystery: '201504',
+    value: 0,
+  },
+  201507: {
+    text: t('backMystery201507Text'),
+    notes: t('backMystery201507Notes'),
+    mystery: '201507',
+    value: 0,
+  },
+  201510: {
+    text: t('backMystery201510Text'),
+    notes: t('backMystery201510Notes'),
+    mystery: '201510',
+    value: 0,
+  },
+};
+
+let eyewear = {
+  201503: {
+    text: t('eyewearMystery201503Text'),
+    notes: t('eyewearMystery201503Notes'),
+    mystery: '201503',
+    value: 0,
+  },
+  201506: {
+    text: t('eyewearMystery201506Text'),
+    notes: t('eyewearMystery201506Notes'),
+    mystery: '201506',
+    value: 0,
+  },
+  201507: {
+    text: t('eyewearMystery201507Text'),
+    notes: t('eyewearMystery201507Notes'),
+    mystery: '201507',
+    value: 0,
+  },
+  301404: {
+    text: t('eyewearMystery301404Text'),
+    notes: t('eyewearMystery301404Notes'),
+    mystery: '301404',
+    value: 0,
+  },
+  301405: {
+    text: t('eyewearMystery301405Text'),
+    notes: t('eyewearMystery301405Notes'),
+    mystery: '301405',
+    value: 0,
+  },
+};
+
+let head = {
+  201402: {
+    text: t('headMystery201402Text'),
+    notes: t('headMystery201402Notes'),
+    mystery: '201402',
+    value: 0,
+  },
+  201405: {
+    text: t('headMystery201405Text'),
+    notes: t('headMystery201405Notes'),
+    mystery: '201405',
+    value: 0,
+  },
+  201406: {
+    text: t('headMystery201406Text'),
+    notes: t('headMystery201406Notes'),
+    mystery: '201406',
+    value: 0,
+  },
+  201407: {
+    text: t('headMystery201407Text'),
+    notes: t('headMystery201407Notes'),
+    mystery: '201407',
+    value: 0,
+  },
+  201408: {
+    text: t('headMystery201408Text'),
+    notes: t('headMystery201408Notes'),
+    mystery: '201408',
+    value: 0,
+  },
+  201411: {
+    text: t('headMystery201411Text'),
+    notes: t('headMystery201411Notes'),
+    mystery: '201411',
+    value: 0,
+  },
+  201412: {
+    text: t('headMystery201412Text'),
+    notes: t('headMystery201412Notes'),
+    mystery: '201412',
+    value: 0,
+  },
+  201501: {
+    text: t('headMystery201501Text'),
+    notes: t('headMystery201501Notes'),
+    mystery: '201501',
+    value: 0,
+  },
+  201505: {
+    text: t('headMystery201505Text'),
+    notes: t('headMystery201505Notes'),
+    mystery: '201505',
+    value: 0,
+  },
+  201508: {
+    text: t('headMystery201508Text'),
+    notes: t('headMystery201508Notes'),
+    mystery: '201508',
+    value: 0,
+  },
+  201509: {
+    text: t('headMystery201509Text'),
+    notes: t('headMystery201509Notes'),
+    mystery: '201509',
+    value: 0,
+  },
+  301404: {
+    text: t('headMystery301404Text'),
+    notes: t('headMystery301404Notes'),
+    mystery: '301404',
+    value: 0,
+  },
+  301405: {
+    text: t('headMystery301405Text'),
+    notes: t('headMystery301405Notes'),
+    mystery: '301405',
+    value: 0,
+  },
+};
+
+let headAccessory = {
+  201403: {
+    text: t('headAccessoryMystery201403Text'),
+    notes: t('headAccessoryMystery201403Notes'),
+    mystery: '201403',
+    value: 0,
+  },
+  201404: {
+    text: t('headAccessoryMystery201404Text'),
+    notes: t('headAccessoryMystery201404Notes'),
+    mystery: '201404',
+    value: 0,
+  },
+  201409: {
+    text: t('headAccessoryMystery201409Text'),
+    notes: t('headAccessoryMystery201409Notes'),
+    mystery: '201409',
+    value: 0,
+  },
+  201502: {
+    text: t('headAccessoryMystery201502Text'),
+    notes: t('headAccessoryMystery201502Notes'),
+    mystery: '201502',
+    value: 0,
+  },
+  201510: {
+    text: t('headAccessoryMystery201510Text'),
+    notes: t('headAccessoryMystery201510Notes'),
+    mystery: '201510',
+    value: 0,
+  },
+  301405: {
+    text: t('headAccessoryMystery301405Text'),
+    notes: t('headAccessoryMystery301405Notes'),
+    mystery: '301405',
+    value: 0,
+  },
+};
+
+let shield = {
+  301405: {
+    text: t('shieldMystery301405Text'),
+    notes: t('shieldMystery301405Notes'),
+    mystery: '301405',
+    value: 0,
+  },
+};
+
+let weapon = {
+  201411: {
+    text: t('weaponMystery201411Text'),
+    notes: t('weaponMystery201411Notes'),
+    mystery: '201411',
+    value: 0,
+  },
+  201502: {
+    text: t('weaponMystery201502Text'),
+    notes: t('weaponMystery201502Notes'),
+    mystery: '201502',
+    value: 0,
+  },
+  201505: {
+    text: t('weaponMystery201505Text'),
+    notes: t('weaponMystery201505Notes'),
+    mystery: '201505',
+    value: 0,
+  },
+  301404: {
+    text: t('weaponMystery301404Text'),
+    notes: t('weaponMystery301404Notes'),
+    mystery: '301404',
+    value: 0,
+  },
+};
+
+let mysterySet = {
+  armor,
+  back,
+  eyewear,
+  head,
+  headAccessory,
+  shield,
+  weapon,
+};
+
+export default mysterySet;
