+++ common/script/content/gear/shield.js
