+++ common/script/content/gear/weapon.js
